# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: ERP-Go API Gateway
  description: |
    跨境电商 ERP 系统 API 网关契约。
    所有请求需经过 API Gateway（端口 8080），Gateway 负责 JWT 鉴权、租户上下文注入和反向代理。
  version: "0.9.0"
  contact:
    name: ERP-Go Team
servers:
  - url: http://localhost:8080
    description: 本地开发
  - url: https://api.erp-go.example.com
    description: 生产环境（示例）
security:
  - BearerAuth: []
paths:
...
  /api/v1/iam/login:
    post:
      summary: 用户登录
      tags: [IAM]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /api/v1/iam/refresh:
    post:
      summary: 刷新令牌
      tags: [IAM]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: 令牌刷新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
  /api/v1/iam/check-permission:
    post:
      summary: 权限检查
      tags: [IAM]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckPermissionRequest"
      responses:
        "200":
          description: 权限检查结果
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT 认证令牌。通过 /api/v1/iam/login 获取。
        请求头: Authorization: Bearer <token>
  schemas:
...
    LoginRequest:
      type: object
      required: [tenant_id, username, password]
      properties:
        tenant_id: { type: string, example: "default" }
        username: { type: string, example: "admin" }
        password: { type: string, example: "admin123" }
    LoginResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: "ok" }
        data:
          type: object
          properties:
            access_token: { type: string }
            refresh_token: { type: string }
            token_type: { type: string, example: "Bearer" }
            expires_in: { type: integer, example: 7200 }
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    CheckPermissionRequest:
      type: object
      required: [user_id, tenant_id, permission_code]
      properties:
        user_id: { type: string }
        tenant_id: { type: string }
        permission_code: { type: string }
...